生徒プロフィール保存のカスタムフック化と軽微な修正、consolelogの削除
