--- components/chat/chatList/ChatNameList.js
@@ -1,17 +1,0 @@
-import ChatName from "../chatRoom/ChatNameBox";
-import Status from "../../common/buttons/StatusButton";
-
-export default function ChatNameList() {
-  return (
-    <div className="h-32 w-[50rem] bg-blue-100 py-2 px-5 rounded">
-      <div className="flex items-center py-2 mb-2">
-        <ChatName />
-        <Status />
-      </div>
-
-      <p className="px-5">
-        よろしくおねがいします！明日の16時からビデオ相談承知しました！
-      </p>
-    </div>
-  );
-}
