--- components/chat/chatRoom/ChatNameBox.js
@@ -1,10 +1,0 @@
-import IconSmall from "../../common/icon/SmallIcon";
-
-export default function ChatName() {
-  return (
-    <div className="flex pl-4 mr-5 items-center">
-      <IconSmall />
-      <p className="ml-4">松丸　慎吾</p>
-    </div>
-  );
-}
