--- components/chat/chatRoom/SendButton.js
@@ -1,9 +1,0 @@
-export default function Send() {
-  return (
-    <>
-      <button className="bg-origin-blue hover:bg-origin-deepBlue text-white px-4 py-1 rounded">
-        送信
-      </button>
-    </>
-  );
-}
