+++ components/student/profileEdit/CancelButton.tsx
@@ -1,8 +1,15 @@
 import Router from "next/router";
+import { useRecoilState } from "recoil";
+import { errorState } from "../../common/atoms";
 
 export default function CancelButton() {
+  const [error, setError] = useRecoilState(errorState);
 
   const handleGoDetailPage = () => {
+    setError({...error, 
+      nameError: "", 
+      gradeError: ""
+    })
     Router.push("/myselfStudentDetail");
   };
